requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
Brotli==1.1.0
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
crispy-bootstrap5==2025.6
cryptography==45.0.4
cssselect2==0.8.0
customtkinter==5.2.2
darkdetect==0.8.0
Django==5.2.1
django-admin-interface==0.30.0
django-admin-rangefilter==0.13.2
django-adminlte3==0.1.6
django-autocomplete-light==3.12.1
django-colorfield==0.14.0
django-crispy-forms==2.4
django-filter==25.1
django-jazzmin==3.0.1
fonttools==4.58.2
html5lib==1.1
idna==3.10
lxml==5.4.0
oscrypto==1.3.0
packaging==25.0
pillow==11.2.1
pycparser==2.22
pydyf==0.11.0
pyHanko==0.29.0
pyhanko-certvalidator==0.27.0
pypdf==5.6.0
pyphen==0.17.2
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-slugify==8.0.4
PyYAML==6.0.2
reportlab==4.4.1
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
svglib==1.5.1
text-unidecode==1.3
tinycss2==1.4.0
tinyhtml5==2.0.0
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.4.0
weasyprint==65.1
webencodings==0.5.1
xhtml2pdf==0.2.17
zopfli==0.2.3.post1
